requirements.txt
streamlit==1.48.1
numpy==2.3.2
pandas==2.3.2
plotly==6.3.0
python-dotenv==1.1.1
alpha_vantage==3.0.0